commented ass4 and 5 and modified ass6
made changes in the makeUnorderedPoly formerly known as (makeRandomPoly) used Collection.shuffle, .removeIf(), and lambda function and made a copy of the terms by newArraylist(terms) it was fun!.<br>
Used enhanced for loops in the makePoly method.<br>
Also got a function that will generate the coefficient form and rearrangements of that it was there where i realised the power of the lambdas! and Collections!<br>
modified the coeffs initialisation loop to enchanced for loop and used Sets instead of map for the while loop to ensure the uniqueness of the options!
modified the solution creator in the Polynomial Class and incorporated the correct answer and index form in it.
Created a termsgen method.
Added a PolyGen function.
created the main function.
removed the null pointer exception in the addition of mts